backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic>=2.10,<3
web3>=6.0,<8
websocket-client>=1.6.0
